+++ Pandora/Paradigm/Primary/Transformer/Tap.hs
@@ -6,7 +6,6 @@ import Pandora.Core.Functor (type (:=))
 import Pandora.Pattern.Semigroupoid ((.))
 import Pandora.Pattern.Category (($), (#))
 import Pandora.Pattern.Functor.Covariant (Covariant ((<$>)), Covariant_ ((-<$>-)))
-import Pandora.Pattern.Functor.Avoidable (Avoidable (empty))
 import Pandora.Pattern.Functor.Pointable (Pointable (point))
 import Pandora.Pattern.Functor.Extractable (Extractable (extract))
 import Pandora.Pattern.Functor.Semimonoidal (Semimonoidal (multiply_))
@@ -36,9 +35,6 @@ instance Covariant t => Covariant (Tap t) where
 
 instance Covariant_ t (->) (->) => Covariant_ (Tap t) (->) (->) where
 	f -<$>- Tap x xs = Tap # f x # f -<$>- xs
-
-instance (Avoidable t, Covariant_ t (->) (->)) => Pointable (Tap t) (->) where
-	point = Tap % empty
 
 instance (Covariant_ t (->) (->)) => Extractable (Tap t) (->) where
 	extract (Tap x _) = x